Tidy up code for parsing OpenAPI
